requirements.txt
Flask==2.3.3
aiohttp==3.8.6
beautifulsoup4==4.12.2
google-generativeai==0.3.2
gunicorn==21.2.0
lxml==4.9.3
requests==2.31.0
python-dotenv==1.0.0
